Написаны классы Test и Fixture, реализованы и пройдены тесты
