--- AirlineCompany/AirlineCompany.Domain/airflyights.csv
@@ -1,10 +1,0 @@
-CodeNumber;DeparturePoint;ArrivalPoint;Departure;Arrive;FlyingTime;PlaneType
-"1000";"Tokio";"Dublin";"2024-03-23 16:39:00";"2024-03-23 21:00:00";"4:21:00";"NG Model 15001"
-"2000";"Rome";"Amsterdam";"2024-09-01 16:35:00";"2024-09-02 01:39:00";"9:04:00";"Panda 202208"
-"3000";"Rome";"Vienna";"2024-11-23 23:50:00";"2024-11-24 12:01:00";"12:11:00";"Panda 202208"
-"4000";"London";"St. Petersburg";"2024-06-26 17:17:00";"2024-06-26 22:39:00";"5:22:00";"NG Model 22009"
-"5000";"Moscow";"Amsterdam";"2024-09-25 23:01:00";"2024-09-26 02:21:00";"3:20:00";"Panda 202015"
-"6000";"Samara";"Dublin";"2024-12-19 04:20:00";"2024-12-19 14:30:00";"10:10:00";"Panda 202208"
-"7000";"Rome";"Dublin";"2024-01-28 19:20:00";"2024-01-29 03:46:00";"8:26:00";"NG Model 15001"
-"8000";"Tokio";"Washington";"2024-06-03 15:43:00";"2024-06-03 19:11:00";"3:28:00";"Panda 202208"
-"9000";"Moscow";"Vienna";"2024-07-13 12:23:00";"2024-07-14 00:20:00";"11:57:00";"Panda 202015"
