--- AirlineCompany/AirlineCompany.Domain/passengers.csv
@@ -1,21 +1,0 @@
-FIO;Passport;Registration;Ticketnumber;SeatNumber;BaggageWeight;CodeFlight
-"Nenashev Ivan Nickolayevich";"7492748492";"1";"30264";"B4";"6.6";"1000"
-"Chumackov Michail Vladimirovich";"837968585";"1";"95613";"A20";"6.3";"8000"
-"Nenashev Ivan Nickolayevich";"7492748492";"1";"55503";"A4";"8.9";"9000"
-"Belyackova Veronicka Sergeevna";"3678678903";"1";"66109";"E2";"2.9";"4000"
-"Belyackova Veronicka Sergeevna";"3678678903";"1";"66431";"E11";"1.0";"8000"
-"Nenashev Ivan Nickolayevich";"7492748492";"1";"34691";"D19";"6.0";"6000"
-"Nenashev Ivan Nickolayevich";"7492748492";"1";"15944";"D12";"8.9";"4000"
-"Chumackov Michail Vladimirovich";"837968585";"1";"64650";"A10";"8.0";"9000"
-"Zamotohina Maria Andreevna";"2875909090";"1";"59641";"A6";"9.3";"5000"
-"Belyackova Veronicka Sergeevna";"3678678903";"1";"65503";"A6";"7.1";"2000"
-"Chumackov Michail Vladimirovich";"837968585";"1";"37572";"E18";"6.5";"1000"
-"Zhuckov Semyon Viktorovich";"5678359284";"1";"85084";"B8";"9.1";"6000"
-"Chumackov Michail Vladimirovich";"837968585";"1";"68293";"D18";"2.5";"6000"
-"Zamotohina Maria Andreevna";"2875909090";"1";"79076";"D9";"9.0";"4000"
-"Zemlyanov Vladimir Borisovich";"1234567890";"1";"67654";"F17";"9.9";"7000"
-"Zamotohina Maria Andreevna";"2875909090";"1";"17438";"E11";"4.3";"8000"
-"Nenashev Ivan Nickolayevich";"7492748492";"1";"83456";"A17";"3.4";"3000"
-"Chumackov Michail Vladimirovich";"837968585";"1";"19030";"F4";"5.9";"2000"
-"Chumackov Michail Vladimirovich";"837968585";"1";"36407";"C6";"1.1";"3000"
-"Nenashev Ivan Nickolayevich";"7492748492";"1";"95201";"E20";"3.1";"7000"
